fix(obfuscate_mail_addr): Add username char limit
Function 'obfuscate_mail_addr' now only hides the part of a mail's
username exceeding a given character limit (default 4) and returns
unchanged mail address if the username length is less than the limit.
Previously it was falsely applying obfuscation indicators (asterisks)
to such addresses. Added better docstring for it.

Minor changes to comments and docs in crud and authutils.